+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+- Patch: Update Datawrapper Switching example with `title` and `ariaLabel` props
+
 ## v0.11.1
 
 - Fix: Correct path for icon exports